+++ flutter/ios/Classes/SentryFlutterPluginApple.swift
@@ -262,6 +262,10 @@ public class SentryFlutterPluginApple: NSObject, FlutterPlugin {
             options.enableOutOfMemoryTracking = enableOutOfMemoryTracking
         }
 
+        if let enableProfiling = arguments["enableProfiling"] as? Bool {
+            options.enableProfiling = enableProfiling
+        }
+
         if let sendClientReports = arguments["sendClientReports"] as? Bool {
             options.sendClientReports = sendClientReports
         }