+++ flutter/lib/src/default_integrations.dart
@@ -364,6 +364,7 @@ class NativeSdkIntegration extends Integration<SentryFlutterOptions> {
         'enableOutOfMemoryTracking': options.enableOutOfMemoryTracking,
         'enableNdkScopeSync': options.enableNdkScopeSync,
         'enableAutoPerformanceTracking': options.enableAutoPerformanceTracking,
+        'enableProfiling': options.enableProfiling,
         'sendClientReports': options.sendClientReports,
       });